Measure Everything in Beats Rather Than Seconds.
This was a major reconstruction, but made a lot of the code cleaner.
Also, it will be a lot easier going forwards to do a lot of what I want.

As an added bonus, I think it sounds better too!
Perhaps slighly more abrupt right now, but that's just a matter of changing the tempo and durations now.

Happy.