Much Better Mixing of Instruments.
This solves a number of problems.
First, the way I was doing it was dumb.
Second, it would lead to things where an instrument would fade out, and suddenly all instruments would jump to a louder volume.
Dumb!

This is better.

Regression: I lost useful logging. I'll have to reinstate that.