Added Probability.
With random apps it's hard to tell if it's working.
Either way, though, I've written it.

I just listened to the various intervals and decided on a probability for each one.
Not very scientific, but that's fine for me.

Also, I made up the algorithm I use to combine the probabilities.
It may be complete garbage.

So. That's that.