Use a Delay and Event Loop to Not Use 100% CPU.
I got tired of having my computer heat up when I was using this.
So, I put in some delay and an event loop to give it a break.

In order to do that I had to Initialize the video subsystem.
That didn't seem to change anything visually for me, but might on other systems.

If this is the case, I may have to back out those changes, or work around them.